Add files via upload
Topic 5.3 Questions:
1) Find the 10 countries with most projects
2) Find the top 10 major project themes (using column 'mjtheme_namecode')
3) In 2. above you will notice that some entries have only the code and the name is missing. Create a dataframe with the missing names filled in.

First, a dictionary of the key, value pairs was created (with the codes representing the keys, and the project names representing the values). Then, using that dictionary, we iterate over the dataframe created in question 2, and any names that are empty strings are filled by using the dictionary. Finally, I demonstrate that the values have been filled by calling the top 11 values (the 122 previously missing values are no longer present).